requirements.txt
absl-py==1.4.0
aiohttp==3.9.5
aiohttp-cors==0.7.0
aiosignal==1.3.1
alabaster==0.7.16
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attridict==0.0.9
attrs==23.1.0
Babel==2.15.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
cachetools==5.3.3
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.4
colorful==0.5.6
comm==0.2.2
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
Cython==3.0.11
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dbus-python==1.2.18
debugpy==1.8.1
decorator==4.4.2
deepsnap==0.2.1
defusedxml==0.7.1
dgl==2.1.0
distlib==0.3.8
distro==1.7.0
dm-tree==0.1.8
docutils==0.21.2
dp-accounting==0.4.3
dtaidistance==2.3.11
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fasteners==0.14.1
fastjsonschema==2.19.1
fedgraph @ file:///home/u1912321/Documents/fedgraph
filelock==3.14.0
Flask==3.0.3
flatbuffers==24.3.25
flwr==1.4.0
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.0
furo==2024.5.6
future==0.18.2
gast==0.5.4
gdown==5.2.0
google-api-core==2.19.0
google-auth==2.30.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
google-vizier==0.1.11
googleapis-common-protos==1.61.0
graphgym==0.4.0
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.64.1
grpcio-tools==1.62.2
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
hydra-core==1.3.2
idna==3.3
imagesize==1.4.1
importlib-metadata==4.6.4
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
iterators==0.0.2
itsdangerous==2.2.0
jax==0.4.14
jaxlib==0.4.14
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kaleido==0.2.1
keras==2.14.0
keyring==23.5.0
kiwisolver==1.4.5
latexcodec==3.0.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
lightning-utilities==0.11.2
linkify-it-py==2.0.3
littleutils==0.2.2
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
memray==1.12.0
mistune==3.0.2
ml-dtypes==0.2.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
notebook_shim==0.2.4
numpy==1.25.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
ogb==1.3.6
olefile==0.46
omegaconf==2.3.0
opencensus==0.11.4
opencensus-context==0.1.3
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.11.0
outdated==0.2.2
overrides==7.7.0
packaging==22.0
pandas==1.3.5
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.4
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.2
plotly==5.24.1
portpicker==1.6.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-spy==0.3.14
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybtex==0.24.0
pybtex-docutils==1.0.3
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==26.0.3
ray==2.24.0
referencing==0.35.1
reportlab==3.6.8
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
scikit-learn==1.5.0
scipy==1.9.3
seaborn==0.13.2
SecretStorage==3.3.1
semantic-version==2.10.0
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.3.7
sphinx-basic-ng==1.0.0b2
sphinx-gallery==0.16.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-bibtex==2.6.2
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==1.4.20
ssh-import-id==5.11
stack-data==0.6.3
sympy==1.12.1
tenacity==9.0.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.14.1
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-model-optimization==0.7.5
tensorflow-probability==0.22.1
tensorflow_federated==0.80.0
tensorflow_privacy==0.9.0
termcolor==2.4.0
terminado==0.18.1
textual==0.66.0
threadpoolctl==3.5.0
tinycss2==1.3.0
tomli==2.0.1
torch==2.3.1
torch-cluster==1.6.3
torch-geometric-temporal==0.54.0
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-spline-conv==1.2.2
torch_geometric==2.5.3
torchaudio==2.3.1
torchdata==0.7.1
torchmetrics==1.4.0.post0
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
triton==2.3.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
uri-template==1.3.0